extra profile controller
